Refactor DashboardViewModel by extracting NotificationsLoader and introducing DispatcherProvider
- Extracted notification logic into `NotificationsLoader`
- Created `DispatcherProvider` interface and `DefaultDispatcherProvider` implementation for injecting dispatchers
- Provided `DispatcherProvider` via `DispatcherModule` in Dagger/Hilt
- Replaced hardcoded `Dispatchers.IO` with `dispatcherProvider.io`
- Added explicit job tracking (`coursesFlowJob`, `teamsFlowJob`, `userDataJob`) in `DashboardViewModel` to safely cancel and manage long-running Flow collections.
