+++ app/src/main/java/org/ole/planet/myplanet/di/DispatcherModule.kt
@@ -1,0 +1,20 @@
+package org.ole.planet.myplanet.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+import org.ole.planet.myplanet.utils.DefaultDispatcherProvider
+import org.ole.planet.myplanet.utils.DispatcherProvider
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DispatcherModule {
+
+    @Provides
+    @Singleton
+    fun provideDispatcherProvider(): DispatcherProvider {
+        return DefaultDispatcherProvider()
+    }
+}
